solutions/instrumentation/metrics/requirements.txt
flask==2.1.2
requests

opentelemetry_sdk==1.11.1
opentelemetry_api==1.11.1
opentelemetry-exporter-otlp-proto-grpc==1.11.1
opentelemetry-exporter-otlp-proto-http==1.11.1
opentelemetry-distro
opentelemetry-instrumentation-flask
